ref(spans): Separate file for SQL scrubbing
